added structure for large numbers

+++ add.h
@@ -8,9 +8,9 @@ const int MAX_SIZE = 2;
 //A struct to hold the numerator, denominator, and whole number of a fraction
 struct fraction
 {
-    int numerator = 0;
-    int denominator = 0;
-    int wholeNumber = 0;
+    unsigned long long numerator = 0;
+    unsigned long long denominator = 0;
+    unsigned long long wholeNumber = 0;
 };
 
 //An enum to hold the operation to be performed
@@ -35,10 +35,10 @@ class FractionList
         int getSize() const;
 };
 
-bool checkFractionParts(char bar, char slash, int denominator);
+bool checkFractionParts(char bar, char slash, unsigned long long denominator);
 void display(FractionList, fraction, int);
 void simplify(fraction& f);
-int GCD(int, int);
-int LCM(int, int);
+unsigned long long GCD(unsigned long long, unsigned long long);
+unsigned long long LCM(unsigned long long, unsigned long long);
 
 #endif
